requirements.txt
streamlit
numpy
Pillow
matplotlib
scikit-image
scipy
opencv-python-headless
